+++ Sources/ABAlbum/Views/AlbumPreviewView.swift
@@ -7,23 +7,39 @@
 
 import SwiftUI
 import Photos
-import PhotosUI
 
 struct AlbumPreviewView: View {
-    
-    var asset: PHAsset
+    @State var currentAssetLocalIdentifier: String?
+    var album: PHAssetCollection
+    var onIndexChange: ((String?) -> Void)?
     
     var body: some View {
         GeometryReader { proxy in
-           AlbumPreviewCellView(asset: asset)
+            ZStack {
+                TabView(selection: $currentAssetLocalIdentifier) {
+                    ForEach(0..<(album.assetsResult?.count ?? 0)) { index in
+                        AlbumPreviewCellView(asset: album.assetsResult!.object(at: index))
+                            .padding(.trailing, 10)
+                            .tag(album.assetsResult?.object(at: index).localIdentifier)
+                    }
+                    .offset(x: 0, y: -proxy.safeAreaInsets.top)
+                    .ignoresSafeArea(.all, edges: .all)
+                }
+                .tabViewStyle(PageTabViewStyle(indexDisplayMode: .never))
+                .offset(x: 0, y: 0)
+            }
+            .ignoresSafeArea()
         }
-        .ignoresSafeArea()
+        .onReceive(currentAssetLocalIdentifier.publisher) { index in
+            guard onIndexChange != nil else { return }
+            onIndexChange!(index)
+        }
         .navigationBarTitleDisplayMode(.inline)
     }
 }
 
 struct AlbumPreviewView_Previews: PreviewProvider {
     static var previews: some View {
-        AlbumPreviewView(asset: PHAsset())
+        AlbumPreviewView(album: PHAssetCollection())
     }
 }